requirements.txt
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blobfile==0.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chumpy==0.70
click==8.0.3
configparser==5.2.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.4.0
freetype-py==2.2.0
gitdb==4.0.9
GitPython==3.1.20
-e git+https://github.com/openai/guided-diffusion.git@27c20a8fab9cb472df5d6bdd6c8d11c8f430b924#egg=guided_diffusion
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.0.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mpi4py==3.1.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
notebook==6.4.6
numpy==1.19.5
open3d==0.11.2
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg-resources==0.0.0
plyfile==0.7.4
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.12.0
pyglet==1.5.21
Pygments==2.10.0
PyOpenGL==3.1.0
pyparsing==3.0.6
pyrender==0.1.45
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
sentry-sdk==1.5.1
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
smplx==0.1.28
subprocess32==3.5.4
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
torch==1.8.2+cu102
torchaudio==0.8.2
torchvision==0.9.2+cu102
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
trimesh==3.9.36
typing-extensions==4.0.1
urllib3==1.26.7
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
xmltodict==0.12.0
yaspin==2.1.0
zipp==3.6.0
